requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
datasets>=2.12.0
peft>=0.4.0

# Image Processing
pillow>=9.0.0

# Data Processing
numpy>=1.21.0
pandas>=1.5.0

# Utilities
tqdm>=4.64.0
safetensors>=0.3.0
huggingface-hub>=0.15.0
